Inlcudes... More headers....

+++ 202/project3/binary_tree.cpp
@@ -57,7 +57,7 @@ bool BST::add(int days_to_race, int train_type, float workouts_per_week, int rac
     char *str_char;
     str_char = strtok(race_history, " ");  // Split list
     while(str_char) {
-        node->add(stof(str_char)); // add to race history list function
+        node->add(atof(str_char)); // add to race history list function
         str_char = strtok(NULL, " ");
     }
